requirements.txt
streamlit>=1.28.0
zai-sdk>=0.2.0
requests>=2.31.0
